Bring back the MarsBotBoard name and expose tagToTrack as a plain record
